requirements.txt
# Core dependencies
openai>=1.12.0
anthropic>=0.19.0
google-generativeai>=0.3.0
pillow>=10.0.0

# Audio processing
sounddevice>=0.4.6
pydub>=0.25.1
numpy>=1.24.0

# UI and Image processing
tkinter  # Usually comes with Python
pygame>=2.5.0
picamera2>=0.3.12  # For Raspberry Pi camera

# Chinese text conversion
opencc>=1.1.1

# Other utilities
python-dotenv>=1.0.0  # For environment variables if needed

# Optional but recommended
requests>=2.31.0  # For HTTP requests
typing-extensions>=4.5.0  # For better type hints
